services/ingest/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
httpx==0.26.0
openai==1.3.7
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
weaviate-client==4.4.0

